deployment/docker/queue/requirements.txt
celery==5.2.7
flower==1.2.0
redis==3.5.3
watchdog==2.1.7
pandas
numpy
amplpyfinance
